Add NRPE deployment, fix epouta and logrotate automation
* Add requirements.yaml and instructions to README

Also included is the required version of Ansible to run this role

* Remove dependencies from meta

* Remove imagebuilder_clouds

Merge it into imagebuilder_clouds_yaml
This also cuts down on a variable for the cloud name
However the service name is kept
just in case you want to call the systemd service something else than the
cloud name.

* Add NRPE deployment

NRPE user is created, perms handled and a config file also

* Add SELinux rules and fix NRPE command

* Add input.json to check.py

* Move log files to be json

* Add logrotate automation

Currently it's set to monthly but weekly is probably also ok

* Move LinuxSE module creation to a file in files/

Now it's created from defined rules within the role so it can
potentially be changed if needed.

* Add ability to skip NRPE deployment

* Fix permissions

* Add more defaults

* Add environment variable and fix input.json format

Input.json format now follows the recommended standard.
Clouds.yaml now has environment variables within itself which get passed
to the systemd service. This ensures that everything can work through a
proxy too.

* Add os_admin_user to defaults